chore(test): de-kollidiere FA-30/39/40 E2E-Specs + härte build-test-inventory.sh [T001023]
Renumber 5 Playwright-Specs auf frische IDs (FA-52..56), da der Dateinamen-
Präfix mit anderen Features kollidierte (Arena-Banner/E-Rechnung/Failure-Loop
teilten sich FA-30; Coaching/LMStudio/Arena-DB teilten sich FA-39; Admin-
Assets/Arena-Spectator teilten sich FA-40):

  - fa-30-arena-banner.spec.ts         → fa-52-arena-banner.spec.ts
  - fa-30-systemtest-failure-loop.spec.ts → fa-53-systemtest-failure-loop.spec.ts
  - fa-39-coaching-sessions.spec.ts    → fa-54-coaching-sessions.spec.ts
  - fa-39-lmstudio-integration.spec.ts → fa-55-lmstudio-integration.spec.ts
  - fa-40-admin-assets.spec.ts         → fa-56-admin-assets.spec.ts

FA-21-invoice-lifecycle nach fa-21-billing gemerged (war Sub-Feature-Split
derselben Spec) und Quelldatei entfernt.

build-test-inventory.sh Dupes-Gate verschärft: pro (id, kind, tier) maximal
1 Eintrag — verhindert künftige ID-Recycling-Bugs, lässt aber local+prod-Tiers
weiter parallel existieren. Verify-Pfad durch Dupe-Injection getestet.

test-inventory.json + repo-index.json regeneriert; openspec Requirement-IDs
(FA-52..56) und source-Links in software-factory.md / admin-cockpit.md /
billing-pipeline.md / llm-local-dev.md / sessions-server.md synchronisiert.

Refs: docs/systemtest-fragebogen.md (FA-30/39/40 = E-Rechnung / Arena-DB /
Arena-Spectator als SSOT; die kollidierenden Files waren nicht im Fragebogen).
