+++ docs/code-quality/repo-index.json
@@ -5,7 +5,7 @@
       "id": "tests",
       "name": "Test suites (BATS + Playwright + vitest)",
       "owner_agent": "bachelorprojekt-test",
-      "file_count": 486,
+      "file_count": 485,
       "files": [
         "tests/.gitignore",
         "tests/batch/batch-gap-analysis.bats",
@@ -68,16 +68,13 @@
         "tests/e2e/specs/fa-18-transcription.spec.ts",
         "tests/e2e/specs/fa-20-finalize.spec.ts",
         "tests/e2e/specs/fa-21-billing.spec.ts",
-        "tests/e2e/specs/fa-21-invoice-lifecycle.spec.ts",
         "tests/e2e/specs/fa-23-vaultwarden.spec.ts",
         "tests/e2e/specs/fa-24-whiteboard.spec.ts",
         "tests/e2e/specs/fa-25-mailpit.spec.ts",
         "tests/e2e/specs/fa-26-bug-report-form.spec.ts",
         "tests/e2e/specs/fa-27-brett.spec.ts",
         "tests/e2e/specs/fa-28-messaging.spec.ts",
-        "tests/e2e/specs/fa-30-arena-banner.spec.ts",
         "tests/e2e/specs/fa-30-einvoice.spec.ts",
-        "tests/e2e/specs/fa-30-systemtest-failure-loop.spec.ts",
         "tests/e2e/specs/fa-32-llm-bge-m3.spec.ts",
         "tests/e2e/specs/fa-33-llm-voyage.spec.ts",
         "tests/e2e/specs/fa-34-llm-strict-fail.spec.ts",
@@ -86,9 +83,6 @@
         "tests/e2e/specs/fa-37-workspace-chat.spec.ts",
         "tests/e2e/specs/fa-38-arena-game-client.spec.ts",
         "tests/e2e/specs/fa-39-arena-db.spec.ts",
-        "tests/e2e/specs/fa-39-coaching-sessions.spec.ts",
-        "tests/e2e/specs/fa-39-lmstudio-integration.spec.ts",
-        "tests/e2e/specs/fa-40-admin-assets.spec.ts",
         "tests/e2e/specs/fa-40-arena-spectator.spec.ts",
         "tests/e2e/specs/fa-41-admin-hub.spec.ts",
         "tests/e2e/specs/fa-42-platform-assets.spec.ts",
@@ -101,6 +95,11 @@
         "tests/e2e/specs/fa-49-factory-observability.spec.ts",
         "tests/e2e/specs/fa-50-request-correlation.spec.ts",
         "tests/e2e/specs/fa-51-sidekick-navigation.spec.ts",
+        "tests/e2e/specs/fa-52-arena-banner.spec.ts",
+        "tests/e2e/specs/fa-53-systemtest-failure-loop.spec.ts",
+        "tests/e2e/specs/fa-54-coaching-sessions.spec.ts",
+        "tests/e2e/specs/fa-55-lmstudio-integration.spec.ts",
+        "tests/e2e/specs/fa-56-admin-assets.spec.ts",
         "tests/e2e/specs/fa-admin-backup-ops.spec.ts",
         "tests/e2e/specs/fa-admin-backup-settings.spec.ts",
         "tests/e2e/specs/fa-admin-billing-system.spec.ts",
